# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for FanFic AI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for a recipient (newest first)
      parameters:
        - name: recipientName
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/unread-count:
    get:
      operationId: getUnreadNotificationCount
      tags: [notifications]
      summary: Count unread notifications for a recipient
      parameters:
        - name: recipientName
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCount"
  /notifications/mark-read:
    post:
      operationId: markNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read for a recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadBody"
      responses:
        "200":
          description: Updated unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCount"
...
  /admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      summary: List moderation reports (newest first)
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [open, hidden, dismissed] }
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }
  /admin/reports/{id}:
    post:
      operationId: adminResolveReport
      tags: [admin]
      summary: Resolve a report (hide content or dismiss)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolveReportBody" }
      responses:
        "200": { description: Updated report, content: { application/json: { schema: { $ref: "#/components/schemas/Report" } } } }
  /reports:
    post:
      operationId: createReport
      tags: [admin]
      summary: Submit a moderation report (any pen name)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReportBody" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Report" } } } }
...
  /stories/{id}/bookmark:
    get:
      operationId: getBookmarkInfo
      tags: [stories]
      summary: Whether the author has bookmarked the story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: authorName
          in: query
          required: false
          schema: { type: string }
      responses:
        "200": { description: Bookmark info, content: { application/json: { schema: { $ref: "#/components/schemas/BookmarkInfo" } } } }
    post:
      operationId: addBookmark
      tags: [stories]
      summary: Add a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookmarkBody" }
      responses:
        "200": { description: Bookmark info, content: { application/json: { schema: { $ref: "#/components/schemas/BookmarkInfo" } } } }
    delete:
      operationId: removeBookmark
      tags: [stories]
      summary: Remove a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: authorName
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: Bookmark info, content: { application/json: { schema: { $ref: "#/components/schemas/BookmarkInfo" } } } }
  /authors/{name}/bookmarks:
    get:
      operationId: listBookmarks
      tags: [stories]
      summary: List an author's bookmarks (with story details)
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Bookmark" }
  /authors/{name}/history:
    get:
      operationId: listReadingHistory
      tags: [stories]
      summary: Reading history for an author (most recent first)
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HistoryEntry" }
  /stories/{id}/progress:
    get:
      operationId: getReadingProgress
      tags: [stories]
      summary: Get a reader's progress on a story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: authorName
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: Progress, content: { application/json: { schema: { $ref: "#/components/schemas/ReadingProgressInfo" } } } }
    post:
      operationId: setReadingProgress
      tags: [stories]
      summary: Save a reader's progress on a story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateReadingProgressBody" }
      responses:
        "200": { description: Updated progress, content: { application/json: { schema: { $ref: "#/components/schemas/ReadingProgressInfo" } } } }
...
components:
  schemas:
...
    Story:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        genre:
          type: string
        artStyle:
          type: string
        lengthSetting:
          type: string
          enum: [short, medium, long]
        seedPrompt:
          type: ["string", "null"]
        fullText:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        characters:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
        authorName:
          type: string
        coAuthors:
          type: array
          items:
            type: string
          default: []
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
      required:
        - id
        - title
        - genre
        - artStyle
        - lengthSetting
        - status
        - coAuthors
        - authorName
        - createdAt
        - updatedAt
        - likeCount
        - commentCount
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        recipientName:
          type: string
        type:
          type: string
          enum: [comment, co_author_chapter, follow, like, repost]
        actorName:
          type: string
        storyId:
          type: ["integer", "null"]
        payload:
          type: ["object", "null"]
          additionalProperties: true
        readAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, recipientName, type, actorName, createdAt]
    UnreadCount:
      type: object
      properties:
        recipientName:
          type: string
        unread:
          type: integer
      required: [recipientName, unread]
    MarkReadBody:
      type: object
      properties:
        recipientName:
          type: string
          minLength: 1
      required: [recipientName]
...
    Bookmark:
      type: object
      properties:
        id: { type: integer }
        authorName: { type: string }
        storyId: { type: integer }
        createdAt: { type: string }
        story: { $ref: "#/components/schemas/Story" }
      required: [id, authorName, storyId, createdAt]
    BookmarkBody:
      type: object
      properties:
        authorName: { type: string, minLength: 1 }
      required: [authorName]
    BookmarkInfo:
      type: object
      properties:
        storyId: { type: integer }
        bookmarked: { type: boolean }
      required: [storyId, bookmarked]
    ReadingProgressInfo:
      type: object
      properties:
        storyId: { type: integer }
        authorName: { type: string }
        progress: { type: integer }
        updatedAt: { type: ["string", "null"] }
      required: [storyId, authorName, progress]
    UpdateReadingProgressBody:
      type: object
      properties:
        authorName: { type: string, minLength: 1 }
        progress: { type: integer, minimum: 0, maximum: 100 }
      required: [authorName, progress]
    HistoryEntry:
      type: object
      properties:
        storyId: { type: integer }
        progress: { type: integer }
        updatedAt: { type: string }
        story: { $ref: "#/components/schemas/Story" }
      required: [storyId, progress, updatedAt, story]
...
    Report:
      type: object
      properties:
        id: { type: integer }
        targetType: { type: string, enum: [story, comment] }
        targetId: { type: integer }
        reporterName: { type: string }
        reason: { type: string }
        status: { type: string, enum: [open, hidden, dismissed] }
        createdAt: { type: string }
        resolvedAt: { type: ["string", "null"] }
        targetPreview: { type: ["string", "null"] }
      required: [id, targetType, targetId, reporterName, reason, status, createdAt]
    CreateReportBody:
      type: object
      properties:
        targetType: { type: string, enum: [story, comment] }
        targetId: { type: integer }
        reporterName: { type: string, minLength: 1 }
        reason: { type: string, maxLength: 500 }
      required: [targetType, targetId, reporterName]
    ResolveReportBody:
      type: object
      properties:
        action: { type: string, enum: [hide, dismiss] }
      required: [action]
...